Implement rather dependency free all GUI prompts :anchor:
