+++ githooks/prompt/gui/gui-darwin.go
@@ -1,0 +1,3 @@
+// +build darwin
+
+package gui
