+++ githooks/prompt/gui/gui-windows.go
@@ -1,0 +1,3 @@
+// +build windows
+
+package gui
