+++ githooks/prompt/show-gui-impl.go
@@ -1,90 +1,11 @@
 package prompt
 
 import (
-	cm "gabyx/githooks/common"
+	"gabyx/githooks/prompt/gui"
 	strs "gabyx/githooks/strings"
-	"strings"
-
-	"github.com/gen2brain/dlgs"
 )
 
-func formatTitle(p *Context) string {
-	return p.log.GetInfoFormatter(false)("Githooks - Git Hook Manager")
-}
-
-func formatTitleQuestion(p *Context) string {
-	return p.log.GetPromptFormatter(false)("Githooks - Git Hook Manager")
-}
-
-func showPromptDialog(
-	p *Context,
-	text string,
-	defaultAnswer string) (string, error) {
-
-	ans, success, err := dlgs.Entry(formatTitle(p), text, defaultAnswer)
-
-	if err != nil {
-		return "", cm.CombineErrors(err, cm.Error("GUI prompt dialog failed."))
-	} else if !success {
-		return "", PromptCanceled
-	}
-
-	return ans, nil
-}
-
-func showPromptOptionDialog(
-	p *Context,
-	question string,
-	defaultAnswerIdx int,
-	options []string,
-	longOptions []string) (string, error) {
-
-	if len(options) == 2 { // nolint: gomnd
-
-		o1 := strings.ToLower(options[0])
-		o2 := strings.ToLower(options[1])
-
-		if (o1 == "y" && o2 == "n") ||
-			(o1 == "n" && o2 == "y") {
-			// This is a normal yes/no prompt
-
-			defaultCancel := (defaultAnswerIdx == 0 && o1 == "n") ||
-				(defaultAnswerIdx == 1 && o2 == "n")
-
-			yes, err := dlgs.Question(formatTitleQuestion(p), question, defaultCancel)
-			if err != nil {
-				return "",
-					cm.CombineErrors(err, cm.Error("GUI option dialog failed."))
-			}
-
-			if yes {
-				return "y", nil
-			}
-
-			return "n", nil
-		}
-	}
-
-	// Make a list dialog
-	selected, success, err := dlgs.List(formatTitle(p), question, longOptions)
-	if err != nil {
-		return "", cm.CombineErrors(err, cm.Error("GUI option dialog failed."))
-	} else if !success {
-		return "", PromptCanceled
-	}
-
-	// Get the chosen answer
-	for i := range longOptions {
-		if selected == longOptions[i] {
-			return strings.ToLower(options[i]), nil
-		}
-	}
-
-	// User has not chosen anything or canceled...
-	return "", nil
-}
-
-func showPromptLoop(
+func showPromptLoopGUI(
 	p *Context,
 	defaultAnswer string,
 	runPrompt func() (string, error),
@@ -140,6 +61,7 @@ func showPromptLoop(
 
 func showPromptOptionsGUI(
 	p *Context,
+	title string,
 	question string,
 	defaultAnswerIdx int,
 	options []string,
@@ -151,26 +73,27 @@ func showPromptOptionsGUI(
 		defaultAnswer = options[defaultAnswerIdx]
 	}
 
-	return showPromptLoop(
+	return showPromptLoopGUI(
 		p,
 		defaultAnswer,
 		func() (string, error) {
-			return showPromptOptionDialog(p, question, defaultAnswerIdx, options, longOptions)
+			return gui.ShowOptionsDialog(title, question, defaultAnswerIdx, options, longOptions)
 		},
 		validator)
 }
 
 func showPromptGUI(
 	p *Context,
+	title string,
 	text string,
 	defaultAnswer string,
 	validator AnswerValidator) (string, error) {
 
-	return showPromptLoop(
+	return showPromptLoopGUI(
 		p,
 		defaultAnswer,
 		func() (string, error) {
-			return showPromptDialog(p, text, defaultAnswer)
+			return gui.ShowPromptDialog(title, text, defaultAnswer)
 		},
 		validator)
 }
