+++ githooks/prompt/show-gui-impl_test.go
@@ -17,16 +17,16 @@ func TestCoverage(t *testing.T) {
 	os.Stdin = nil
 	promptCtx, _ := CreateContext(log, &cm.ExecContext{}, nil, true, false)
 
-	ans, err := promptCtx.ShowPrompt("Enter a string:", "This is the default string", ValidatorAnswerNotEmpty)
+	ans, err := promptCtx.ShowPrompt("Enter a string sssssssssss:", "This is the default string", ValidatorAnswerNotEmpty)
 	log.InfoF("Answer: '%s'", ans)
 	log.AssertNoErrorF(err, "Error occurred.")
 
-	ans, err = promptCtx.ShowPromptOptions("Choose string:", "(Yes/no)", "Y/n", "Yes", "No")
+	ans, err = promptCtx.ShowPromptOptions("Choose string      ss       ssssssssssssss s s s asd asdfl kjj sdlfök jsaölkdf jölaskdjf lökasjd flökjsa döfl  s:", "(Yes/no)", "Y/n", "Yes", "No")
 	log.InfoF("Answer: '%s'", ans)
 	log.AssertNoErrorF(err, "Error occurred.")
 
 	ans, err = promptCtx.ShowPromptOptions(
-		"This string?", "(Yes/no/skip/skip all)", "Y/n/s/a", "Yes", "No", "Skip", "Skip All")
+		"This string sssssssssss s s as sd asd asd asd\nasd asd asd s              asd asd asd asd asd asd?", "(Yes/no/skip/skip all)", "Y/n/s/a", "Yes", "No", "Skip", "Skip All")
 	log.InfoF("Answer: '%s'", ans)
 	log.AssertNoErrorF(err, "Error occurred.")
